requirements.txt
opencv-python
Pillow
numpy
scenedetect[opencv]
torch
git+https://github.com/openai/CLIP.git
chromadb
insightface
onnxruntime
scikit-learn
google-generativeai
anthropic
openai
datasets
vllm
